author action logs

+++ Data/Services/AuthorServices.cs
@@ -1,3 +1,4 @@
+using Data.Eunumerators;
 using Data.Exceptions;
 using Data.Models;
 using Microsoft.EntityFrameworkCore;
@@ -16,10 +17,12 @@ public interface IAuthorService
 public class AuthorServices : IAuthorService
 {
     private readonly RwaContext _context;
+    private readonly ILogService _logService;
     
-    public AuthorServices(RwaContext context)
+    public AuthorServices(RwaContext context, ILogService logService)
     {
         _context = context;
+        _logService = logService;
     }
     
     public async Task<Author> Get(int id)
@@ -29,16 +32,23 @@ public class AuthorServices : IAuthorService
             .FirstOrDefaultAsync(a => a.IdAuthor == id);
         if (author == null)
         {
+            _logService.Create("Failed to fetch autor with ID {id}, because there is none with the same ID", Importance.Low);
             throw new NotFoundException($"Author with id {id} not found");
         }
-
+        _logService.Create("Author with ID {id} fetched successfully", Importance.Low);
         return author;
     }
 
     public async Task Delete(int id)
     {
         Author author = await _context.Authors.FirstOrDefaultAsync(a => a.IdAuthor == id);
+        if (author == null)
+        {
+            _logService.Create("Cannot delete author with ID {id}, because there is none with the same ID", Importance.Low);
+            throw new NotFoundException($"author with id {id} not found");
+        }
         _context.Authors.Remove(author);
+        _logService.Create("Author with ID {id} successfully deleted", Importance.High);
         await _context.SaveChangesAsync();
     }
 
@@ -47,12 +57,14 @@ public class AuthorServices : IAuthorService
         Author original = await _context.Authors.FirstOrDefaultAsync(a => a.IdAuthor == id);
         if (original == null)
         {
+            _logService.Create("Cannot update author with ID {id}, because there is none with the same ID", Importance.Low);
             throw new NotFoundException($"author with id {id} not found");
         }
 
         original.AuthorName = author.AuthorName;
         
         _context.Authors.Update(original);
+        _logService.Create("Author with ID {id} auccessfully updated", Importance.Medium);
         await _context.SaveChangesAsync();
     }
 
@@ -61,10 +73,12 @@ public class AuthorServices : IAuthorService
         Author? author = await _context.Authors.FirstOrDefaultAsync(a => a.IdAuthor == newAuthor.IdAuthor);
         if (author != null)
         {
+            _logService.Create($"Failed to add new author, because author with ID {newAuthor.IdAuthor} already existst", Importance.Low);
             throw new AlreadyExistsException($"Author with {newAuthor.IdAuthor} already exists");
         }
 
         await _context.Authors.AddAsync(newAuthor);
+        _logService.Create($"Author with ID {newAuthor.IdAuthor} successfully added", Importance.Low);
         await _context.SaveChangesAsync();
     }
